+++ src/services/Products.php
@@ -12,9 +12,11 @@ use craft\helpers\Json;
 use craft\helpers\ProjectConfig;
 use craft\helpers\StringHelper;
 use craft\models\FieldLayout;
+use craft\shopify\collections\VariantCollection;
 use craft\shopify\db\Table;
 use craft\shopify\elements\Product;
 use craft\shopify\events\ShopifyProductSyncEvent;
+use craft\shopify\models\Variant;
 use craft\shopify\Plugin;
 use craft\shopify\records\ShopifyData;
 use GraphQL\QueryBuilder\QueryBuilder;
@@ -289,15 +291,51 @@ class Products extends Component
      */
     public function eagerLoadVariantsForProducts(array $products): array
     {
-        return $this->_eagerLoadTypeOnProducts($products, 'ProductVariant', function($product, $rows) {
-            $product->setVariants(array_column($rows, 'data'));
+        $variantIds = [];
+        $variantsByProductId = [];
+        $return = $this->_eagerLoadTypeOnProducts($products, 'ProductVariant', function($product, $rows) use (&$variantsByProductId, &$variantIds) {
+            foreach ($rows as $row) {
+                $variantIds[] = $row->shopifyId;
+            }
+
+            $variantsByProductId[$product->shopifyGid] = $rows;
         });
+
+        // If we are eager loading the variants, for best performance we should also eager load the metafields on the variants
+        $metafieldsData = collect();
+        if (!empty($variantIds)) {
+            $metafieldsData = Plugin::getInstance()
+                ->getApi()
+                ->getShopifyDataByType('Metafield', $variantIds, true)
+                ->groupBy('parentId');
+        }
+
+        foreach ($return as $product) {
+            $variants = VariantCollection::make($variantsByProductId[$product->shopifyGid]);
+
+            if ($metafieldsData->isNotEmpty()) {
+                $variants?->map(function(Variant$variant) use ($metafieldsData) {
+                    $metafields = $metafieldsData->get($variant->shopifyId);
+                    if (!empty($metafields)) {
+                        $variant->setMetafields(collect($metafields)->mapWithKeys(function($d) {
+                            return [
+                                $d->data['key'] => Json::decodeIfJson($d->data['value']),
+                            ];
+                        })->all());
+                    }
+                });
+            }
+
+            $product->setMetafields($variants);
+        }
+
+        return $return;
     }
 
     /**
      * @param array|Product[] $products
      * @param string $type
-     * @param callable $callback
+     * @param callable(Product, ShopifyData[]): void $callback
      * @return array
      * @throws InvalidConfigException
      */